Revert "Fix concurrent delete failure"
This reverts commit bd3203ff0567c97529675ec716ba7d963d02fea0.

+++ model/jpa/src/main/java/org/keycloak/models/jpa/JpaUserProvider.java
@@ -165,10 +165,16 @@ public class JpaUserProvider implements UserProvider.Streams, UserCredentialStor
         em.createNamedQuery("deleteFederatedIdentityByUser").setParameter("user", user).executeUpdate();
         em.createNamedQuery("deleteUserConsentClientScopesByUser").setParameter("user", user).executeUpdate();
         em.createNamedQuery("deleteUserConsentsByUser").setParameter("user", user).executeUpdate();
-        em.remove(user);
         em.flush();
         // not sure why i have to do a clear() here.  I was getting some messed up errors that Hibernate couldn't
         // un-delete the UserEntity.
+        //em.clear();
+        user = em.find(UserEntity.class, id, LockModeType.PESSIMISTIC_WRITE);
+        if (user != null) {
+            em.remove(user);
+        }
+
+        em.flush();
     }
 
     @Override
@@ -500,9 +506,9 @@ public class JpaUserProvider implements UserProvider.Streams, UserCredentialStor
 
     @Override
     public Stream<UserModel> getGroupMembersStream(RealmModel realm, GroupModel group) {
-        TypedQuery<String> query = em.createNamedQuery("groupMembership", String.class);
+        TypedQuery<UserEntity> query = em.createNamedQuery("groupMembership", UserEntity.class);
         query.setParameter("groupId", group.getId());
-        return closing(query.getResultStream().map(id -> session.users().getUserById(id, realm)));
+        return closing(query.getResultStream().map(entity -> new UserAdapter(session, realm, em, entity)));
     }
 
     @Override
@@ -520,7 +526,7 @@ public class JpaUserProvider implements UserProvider.Streams, UserCredentialStor
 
     @Override
     public UserModel getUserById(String id, RealmModel realm) {
-        UserEntity userEntity = em.find(UserEntity.class, id, LockModeType.PESSIMISTIC_WRITE);
+        UserEntity userEntity = em.find(UserEntity.class, id);
         if (userEntity == null || !realm.getId().equals(userEntity.getRealmId())) return null;
         return new UserAdapter(session, realm, em, userEntity);
     }
